refactor(http): organize endpoints, REST DTOs and their tests by domain
Endpoints/ and Data/Api/ gain one folder per domain (Accounts, Admin,
Auth, Characters, Images, Maps, Players, Version), namespaces aligned
to paths per the code convention, and the endpoint tests mirror the
production layout under Http/Endpoints.
