+++ src/Moongate.Http.Plugin/Data/Api/Auth/LoginRequest.cs
@@ -1,0 +1,4 @@
+namespace Moongate.Http.Plugin.Data.Api.Auth;
+
+/// <summary>Credentials posted to the login endpoint.</summary>
+public sealed record LoginRequest(string Username, string Password);
