+++ src/Moongate.Http.Plugin/Data/Api/Players/PlayerMeResponse.cs
@@ -1,0 +1,4 @@
+namespace Moongate.Http.Plugin.Data.Api.Players;
+
+/// <summary>Who the caller's token says they are. Deliberately not their characters — see PlayerEndpoints.</summary>
+public sealed record PlayerMeResponse(string Username, string Level);
