+++ src/Moongate.Http.Plugin/Data/Api/Version/VersionResponse.cs
@@ -1,0 +1,4 @@
+namespace Moongate.Http.Plugin.Data.Api.Version;
+
+/// <summary>What the shard calls itself and which build it runs.</summary>
+public sealed record VersionResponse(string ShardName, string Version);
