--- src/Moongate.Http.Plugin/Endpoints/PlayerEndpoints.cs
@@ -1,36 +1,0 @@
-using System.Security.Claims;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Routing;
-using Moongate.Http.Plugin.Interfaces;
-using Moongate.Http.Plugin.Services;
-using Moongate.Http.Plugin.Data.Api;
-
-namespace Moongate.Http.Plugin.Endpoints;
-
-/// <summary>What an authenticated account may ask about itself.</summary>
-public sealed class PlayerEndpoints : IApiEndpointRegistration
-{
-    public void Register(IEndpointRouteBuilder routes)
-    {
-        routes.MapGet("/api/v1/player/me", Me)
-              .WithName("GetPlayerMe")
-              .WithTags("player")
-              .RequireAuthorization(HttpServerService.PlayerPolicy);
-    }
-
-    /// <summary>Reports the account the bearer token belongs to.</summary>
-    /// <remarks>
-    /// Username and level, read straight from the token — useful for confirming a token is still valid
-    /// and what it grants. It does not list the account's characters.
-    /// </remarks>
-    // Deliberately no character list: that would read the mobile store, which is single-writer on the game
-    // loop, and drag loop affinity into a probe endpoint.
-    private static IResult Me(ClaimsPrincipal user)
-        => Results.Ok(
-            new PlayerMeResponse(
-                user.FindFirstValue(ClaimTypes.Name) ?? string.Empty,
-                user.FindFirstValue(ClaimTypes.Role) ?? string.Empty
-            )
-        );
-}
