+++ src/Moongate.Http.Plugin/MoongateHttpPlugin.cs
@@ -1,6 +1,13 @@
 using DryIoc;
 using Moongate.Http.Plugin.Data.Config;
-using Moongate.Http.Plugin.Endpoints;
+using Moongate.Http.Plugin.Endpoints.Accounts;
+using Moongate.Http.Plugin.Endpoints.Admin;
+using Moongate.Http.Plugin.Endpoints.Auth;
+using Moongate.Http.Plugin.Endpoints.Characters;
+using Moongate.Http.Plugin.Endpoints.Images;
+using Moongate.Http.Plugin.Endpoints.Maps;
+using Moongate.Http.Plugin.Endpoints.Players;
+using Moongate.Http.Plugin.Endpoints.Version;
 using Moongate.Http.Plugin.Extensions;
 using Moongate.Http.Plugin.Interfaces;
 using Moongate.Http.Plugin.Services;
