+++ tests/Moongate.Tests/Http/MoongateHttpPluginTests.cs
@@ -1,7 +1,14 @@
 using DryIoc;
 using Moongate.Http.Plugin;
 using Moongate.Http.Plugin.Data.Config;
-using Moongate.Http.Plugin.Endpoints;
+using Moongate.Http.Plugin.Endpoints.Accounts;
+using Moongate.Http.Plugin.Endpoints.Admin;
+using Moongate.Http.Plugin.Endpoints.Auth;
+using Moongate.Http.Plugin.Endpoints.Characters;
+using Moongate.Http.Plugin.Endpoints.Images;
+using Moongate.Http.Plugin.Endpoints.Maps;
+using Moongate.Http.Plugin.Endpoints.Players;
+using Moongate.Http.Plugin.Endpoints.Version;
 using Moongate.Http.Plugin.Interfaces;
 using Moongate.Http.Plugin.Services;
 using Moongate.Ultima.Catalog;
