+++ tests/Moongate.Tests/Support/TestApiServer.cs
@@ -8,7 +8,12 @@ using Moongate.Http.Plugin.Data.Config;
 using Moongate.Http.Plugin.Interfaces;
 using Moongate.Http.Plugin.Services;
 using Moongate.Server.Abstractions.Data.Config;
-using Moongate.Http.Plugin.Endpoints;
+using Moongate.Http.Plugin.Endpoints.Accounts;
+using Moongate.Http.Plugin.Endpoints.Admin;
+using Moongate.Http.Plugin.Endpoints.Auth;
+using Moongate.Http.Plugin.Endpoints.Characters;
+using Moongate.Http.Plugin.Endpoints.Players;
+using Moongate.Http.Plugin.Endpoints.Version;
 using Moongate.Server.Abstractions.Interfaces.Accounts;
 using Moongate.Server.Services.Accounts;
 using SquidStd.Core.Interfaces.Config;
